CQ-4301935 Refactor PortalLister interface
- New interface is meant to be used via inheritance and not composition

@review @rismehta @sakshi-arora1
