+++ bundles/core/src/main/java/com/adobe/cq/forms/core/components/internal/models/v1/formsportal/PortalListerImpl.java
@@ -15,48 +15,26 @@
  ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~*/
 package com.adobe.cq.forms.core.components.internal.models.v1.formsportal;
 
-import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashMap;
-import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 
 import javax.inject.Inject;
-import javax.inject.Named;
-import javax.jcr.Session;
 
 import org.apache.commons.lang3.StringUtils;
-import org.apache.jackrabbit.vault.util.JcrConstants;
 import org.apache.sling.api.SlingHttpServletRequest;
-import org.apache.sling.api.request.RequestParameter;
-import org.apache.sling.api.request.RequestParameterMap;
-import org.apache.sling.api.resource.LoginException;
-import org.apache.sling.api.resource.Resource;
-import org.apache.sling.api.resource.ResourceResolver;
-import org.apache.sling.api.resource.ResourceResolverFactory;
-import org.apache.sling.api.resource.ValueMap;
 import org.apache.sling.models.annotations.Default;
 import org.apache.sling.models.annotations.DefaultInjectionStrategy;
 import org.apache.sling.models.annotations.Model;
 import org.apache.sling.models.annotations.Required;
-import org.apache.sling.models.annotations.injectorspecific.ChildResource;
 import org.apache.sling.models.annotations.injectorspecific.InjectionStrategy;
-import org.apache.sling.models.annotations.injectorspecific.OSGiService;
 import org.apache.sling.models.annotations.injectorspecific.Self;
 import org.apache.sling.models.annotations.injectorspecific.ValueMapValue;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import com.adobe.cq.export.json.ComponentExporter;
 import com.adobe.cq.forms.core.components.internal.models.v1.AbstractComponentImpl;
 import com.adobe.cq.forms.core.components.models.formsportal.PortalLister;
-import com.day.cq.i18n.I18n;
-import com.day.cq.search.PredicateGroup;
-import com.day.cq.search.Query;
-import com.day.cq.search.QueryBuilder;
-import com.day.cq.search.result.SearchResult;
-import com.fasterxml.jackson.annotation.JsonIgnore;
 
 @Model(
     adaptables = SlingHttpServletRequest.class,
@@ -66,335 +44,89 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 public class PortalListerImpl extends AbstractComponentImpl implements PortalLister {
     public static final String RESOURCE_TYPE = "core/fd/components/formsportal/portallister/v1/portallister";
 
-    private static final String QB_GROUP = "_group.";
-    private static final String QB_PATH = "_path";
-    private static final String QB_FULLTEXT = "_fulltext";
-    private static final String QB_FULLTEXT_RELPATH = "_fulltext.relPath";
-    private static final String DEFAULT_PATH = "/content/dam/formsanddocuments";
-    private static final Logger LOGGER = LoggerFactory.getLogger(PortalListerImpl.class);
-    private static final String METADATA_NODE_PATH = JcrConstants.JCR_CONTENT + "/metadata";
-    private static final String PN_CHILD_ASSETFOLDERS = "assetFolders";
-    private static final String PN_CHILD_ASSETSOURCES = "assetSource";
-    private static final String THUMBNAIL_PATH_SUFFIX = "/jcr:content/renditions/cq5dam.thumbnail.319.319.png";
-    private static final Map<String, QueryStrategy> queryStrategies = new HashMap<>();
-    static {
-        queryStrategies.put("searchText", new QueryStrategy() {
-            public void buildQuery(RequestParameter[] params, int predicateID, Map<String, String> queryMap) {
-                int counter = 0;
-                int paramCount = params.length;
-                while (paramCount-- > 0) {
-                    String text = params[counter].getString();
-                    if (StringUtils.isNotBlank(text)) {
-                        queryMap.put(predicateID + QB_FULLTEXT, text);
-                        queryMap.put(predicateID + QB_FULLTEXT_RELPATH, METADATA_NODE_PATH);
-                        predicateID++;
-                    }
-                    counter++;
-                }
-            }
-        });
+    @ValueMapValue(injectionStrategy = InjectionStrategy.OPTIONAL)
+    @Inject
+    private String title;
 
-        queryStrategies.put("title", new QueryStrategy() {
-            public void buildQuery(RequestParameter[] params, int predicateID, Map<String, String> queryMap) {
-                String title = params[0].getString();
-                queryMap.put(predicateID + QB_GROUP + "0" + QB_FULLTEXT, title);
-                queryMap.put(predicateID + QB_GROUP + "0" + QB_FULLTEXT_RELPATH, METADATA_NODE_PATH + "/@title");
-                queryMap.put(predicateID + QB_GROUP + "1" + QB_FULLTEXT, title);
-                queryMap.put(predicateID + QB_GROUP + "1" + QB_FULLTEXT_RELPATH, METADATA_NODE_PATH + "/@dc:title");
-                queryMap.put(predicateID + QB_GROUP + "p.or", "true");
-            }
-        });
+    @ValueMapValue(injectionStrategy = InjectionStrategy.OPTIONAL)
+    @Inject
+    private String layout;
 
-        queryStrategies.put("description", new QueryStrategy() {
-            public void buildQuery(RequestParameter[] params, int predicateID, Map<String, String> queryMap) {
-                String description = params[0].getString();
-                queryMap.put(predicateID + QB_GROUP + "0" + QB_FULLTEXT, description);
-                queryMap.put(predicateID + QB_GROUP + "0" + QB_FULLTEXT_RELPATH, METADATA_NODE_PATH + "/@description");
-                queryMap.put(predicateID + QB_GROUP + "1" + QB_FULLTEXT, description);
-                queryMap.put(predicateID + QB_GROUP + "1" + QB_FULLTEXT_RELPATH, METADATA_NODE_PATH + "/@dc:description");
-                queryMap.put(predicateID + QB_GROUP + "p.or", "true");
-            }
-        });
+    @ValueMapValue
+    @Inject
+    @Default(intValues = 8)
+    private Integer limit;
 
-        queryStrategies.put("tags", new QueryStrategy() {
-            public void buildQuery(RequestParameter[] params, int predicateID, Map<String, String> queryMap) {
-                int tagID = 0;
-                int paramCount = params.length;
-                while (paramCount-- > 0) {
-                    queryMap.put(predicateID + "_tagid", params[tagID].getString());
-                    queryMap.put(predicateID + "_tagid.property", METADATA_NODE_PATH + "/cq:tags");
-                    predicateID++;
-                    tagID++;
-                }
-            }
-        });
-
-        queryStrategies.put("orderby", new QueryStrategy() {
-            public void buildQuery(RequestParameter[] params, int predicateID, Map<String, String> queryMap) {
-                String orderByValue = params[0].getString();
-                queryMap.put("orderby", "@" + METADATA_NODE_PATH + "/" + orderByValue);
-            }
-        });
-
-        queryStrategies.put("sort", new QueryStrategy() {
-            public void buildQuery(RequestParameter[] params, int predicateID, Map<String, String> queryMap) {
-                queryMap.put("orderby.sort", params[0].getString());
-            }
-        });
-
-        queryStrategies.put("offset", new QueryStrategy() {
-            public void buildQuery(RequestParameter[] params, int predicateID, Map<String, String> queryMap) {
-                queryMap.put("p.offset", params[0].getString());
-            }
-        });
+    @Override
+    public Integer getLimit() {
+        return limit;
     }
-
-    @OSGiService
-    private QueryBuilder queryBuilder;
-
-    @OSGiService
-    private ResourceResolverFactory resourceResolverFactory;
 
     @Self
     @Required
     private SlingHttpServletRequest request;
 
-    @ChildResource(injectionStrategy = InjectionStrategy.OPTIONAL)
-    @Named(PN_CHILD_ASSETFOLDERS)
-    @Inject
-    private List<Resource> assetFolders;
-
-    @ChildResource(injectionStrategy = InjectionStrategy.OPTIONAL)
-    @Named(PN_CHILD_ASSETSOURCES)
-    @Inject
-    private List<Resource> assetSources;
-
-    private List<Resource> defaultAssetSources;
-
-    @ValueMapValue
-    @Inject
-    @Default(longValues = 8)
-    private long limit;
-
-    private String htmlTooltip = "";
-    private String pdfTooltip = "";
-    private Long nextOffset = 0l;
-
-    private String getAssetType() {
-        // Decide on how to expose this method in authoring?
-        return "dam:Asset";
-    }
-
-    private String getDataSource() {
-        // Decide on how to expose this method in authoring?
-        return "AEM";
-    }
-
-    @Override
-    public List<PortalLister.Item> getItemList() {
-        // call implementation depending on data source
-        List<PortalLister.Item> result = null;
-        if ("AEM".equals(getDataSource())) {
-            try (ResourceResolver resourceResolver = resourceResolverFactory.getServiceResourceResolver(null)) {
-                result = fetchViaQueryBuilder(resourceResolver);
-            } catch (LoginException e) {
-                e.printStackTrace();
-            }
+    protected Integer getOffset() {
+        String offsetParam = request.getParameter("offset");
+        if (StringUtils.isNumeric(offsetParam)) {
+            return Integer.valueOf(offsetParam);
         }
-        return result;
+        return 0;
     }
 
-    @Override
-    public Long getNextOffset() {
+    protected Integer getNextOffset(Integer resultLength) {
+        Integer nextOffset = getOffset();
+        Integer limit = getLimit();
+        if (resultLength < limit) {
+            nextOffset = -1;
+        } else {
+            // in the case fetched results are equal to limit, we don't know if there are more or not
+            // determining it lazily
+            nextOffset = nextOffset + resultLength;
+        }
         return nextOffset;
     }
 
     @Override
-    public Long getLimit() {
-        String limitParam = request.getParameter("limit");
-        if (StringUtils.isNumeric(limitParam)) {
-            return Long.valueOf(limitParam);
-        }
-        return limit;
+    public Map<String, Object> getSearchResults() {
+        List<PortalLister.Item> results = getItemList();
+        Integer offset = getOffset();
+        Integer nextOffset = getNextOffset(results.size());
+        Map<String, Object> jacksonMapping = new HashMap<>();
+        jacksonMapping.put("data", results);
+        jacksonMapping.put("nextOffset", nextOffset);
+        jacksonMapping.put("offset", offset);
+        return jacksonMapping;
+    }
+
+    /**
+     * Function to be implemented by child classes for returning item list
+     */
+    protected List<PortalLister.Item> getItemList() {
+        return Collections.EMPTY_LIST;
+    }
+
+    protected String defaultTitle() {
+        return StringUtils.EMPTY;
     }
 
     @Override
-    @JsonIgnore
-    public List<Resource> getAssetSources() {
-        if (assetSources != null) {
-            return assetSources;
-        } else if (defaultAssetSources != null) {
-            return defaultAssetSources;
+    public String getTitle() {
+        if (title == null) {
+            return defaultTitle();
         }
-        return Collections.emptyList();
+        return title;
     }
 
     @Override
-    @JsonIgnore
-    public void setDefaultAssetSources(List<Resource> assetSources) {
-        this.defaultAssetSources = assetSources;
+    public String getLayout() {
+        if (StringUtils.isEmpty(layout)) {
+            return PortalLister.LayoutType.CARD;
+        }
+        return layout;
     }
 
-    private int buildAssetSourcesQuery(int predicateID, Map<String, String> queryMap) {
-        List<Resource> assetSourcesOrDefault = getAssetSources();
-        if (assetSourcesOrDefault != null) {
-            I18n i18n = new I18n(request);
-            int counter = 1;
-            for (Resource source : assetSourcesOrDefault) {
-                ValueMap assetSource = source.getValueMap();
-                String renderType = assetSource.get("type", String.class);
-                String typeValue = null;
-                String renderKey = null;
-                if (StringUtils.isBlank(renderType)) {
-                    continue;
-                } else if (renderType.equals("Adaptive Forms")) {
-                    typeValue = "guide";
-                    renderKey = "HTML";
-                    htmlTooltip = i18n.get(assetSource.get("htmlTooltip", String.class));
-                } else if (renderType.equals("PDF Forms")) {
-                    typeValue = "pdfForm";
-                    renderKey = "PDF";
-                    pdfTooltip = i18n.get(assetSource.get("pdfTooltip", String.class));
-                }
-
-                String typePrefix = "group." + predicateID + QB_GROUP + counter + "_property";
-                queryMap.put(typePrefix, "./jcr:content/type");
-                queryMap.put(typePrefix + ".value", typeValue);
-                queryMap.put(typePrefix + ".operation", "equals");
-                counter++;
-
-                typePrefix = "group." + predicateID + QB_GROUP + counter + "_property";
-                queryMap.put(typePrefix, "./jcr:content/metadata/allowedRenderFormat");
-                queryMap.put(typePrefix + ".0_value", "BOTH");      // this is ANDed with set renderKey
-                queryMap.put(typePrefix + ".1_value", renderKey);
-                queryMap.put(typePrefix + ".operation", "equals");
-
-                counter++;
-                predicateID++;
-            }
-            queryMap.put("group.p.or", "true");
-        }
-        return predicateID;
-    }
-
-    private int buildAssetFolderQuery(int predicateID, Map<String, String> queryMap) {
-        if (assetFolders != null) {
-            int counter = 0;
-            for (Resource source : assetFolders) {
-                ValueMap assetFolder = source.getValueMap();
-                String folderPath = assetFolder.get("folder", String.class);
-                if (StringUtils.isNotBlank(folderPath)) {
-                    queryMap.put(predicateID + QB_GROUP + counter + QB_PATH, folderPath);
-                    counter++;
-                }
-            }
-        } else {
-            // default search path
-            queryMap.put(predicateID + QB_GROUP + "0" + QB_PATH, DEFAULT_PATH);
-        }
-        queryMap.put(predicateID + "_group.p.or", Boolean.TRUE.toString());
-        return predicateID + 1;
-    }
-
-    protected List<PortalLister.Item> fetchViaQueryBuilder(ResourceResolver resourceResolver) {
-        RequestParameterMap parameterMap = request.getRequestParameterMap();
-        Session session = null;
-        // service resource resolver has fd-user access, ref OSGi configuration file
-        session = resourceResolver.adaptTo(Session.class);
-        List<PortalLister.Item> resultMap = new ArrayList<>();
-        Map<String, String> queryMap = new HashMap<>();
-
-        queryMap.put("type", getAssetType());
-        queryMap.put("p.limit", getLimit().toString());
-
-        int predicateID = 0;    // used across all parameters
-        for (Map.Entry<String, RequestParameter[]> entry : parameterMap.entrySet()) {
-            QueryStrategy queryStrategy = queryStrategies.get(entry.getKey());
-            if (queryStrategy != null) {
-                queryStrategy.buildQuery(entry.getValue(), predicateID, queryMap);
-            }
-            predicateID++;
-        }
-
-        predicateID = buildAssetSourcesQuery(predicateID, queryMap);
-        predicateID = buildAssetFolderQuery(predicateID, queryMap);
-
-        PredicateGroup predicateGroup = PredicateGroup.create(queryMap);
-        Query query = queryBuilder.createQuery(predicateGroup, session);
-        SearchResult searchResult = query.getResult();
-
-        long totalMatches = searchResult.getTotalMatches();
-        long currentHits = searchResult.getHits().size();
-        long startIndex = searchResult.getStartIndex();
-        if (startIndex + currentHits < totalMatches) {
-            nextOffset = startIndex + currentHits;
-        }
-
-        // Query builder has a leaking resource resolver, so the following work around is required.
-        ResourceResolver leakingResourceResolver = null;
-        try {
-            Iterator<Resource> resourceIterator = searchResult.getResources();
-            while (resourceIterator.hasNext()) {
-                Resource resource = resourceIterator.next();
-                if (leakingResourceResolver == null) {
-                    // Get a reference to QB's leaking resource resolver
-                    leakingResourceResolver = resource.getResourceResolver();
-                }
-                resultMap.add(fetchResourceProperties(resource));
-            }
-        } finally {
-            if (leakingResourceResolver != null) {
-                // Always close the leaking query builder resource resolver
-                leakingResourceResolver.close();
-            }
-        }
-
-        return resultMap;
-    }
-
-    private PortalLister.Item fetchResourceProperties(Resource r) {
-        ValueMap valueMap = r.getValueMap();
-        ResourceResolver resolver = r.getResourceResolver();
-        ValueMap metaDataMap = null;
-        if (r.getChild(METADATA_NODE_PATH) != null) {
-            metaDataMap = r.getChild(METADATA_NODE_PATH).getValueMap();
-        }
-        String title = null;
-        String description = null;
-        String path = r.getPath() + "/" + JcrConstants.JCR_CONTENT + "?wcmmode=disabled";
-        String tooltip = "";
-        String thubmnail = r.getPath() + THUMBNAIL_PATH_SUFFIX;
-
-        if (metaDataMap != null && metaDataMap.containsKey("title")) {
-            title = metaDataMap.get("title", String.class);
-        } else if (valueMap.containsKey(JcrConstants.JCR_TITLE)) {
-            title = valueMap.get(JcrConstants.JCR_TITLE, String.class);
-        }
-        if (metaDataMap != null && metaDataMap.containsKey("dc:description")) {
-            description = metaDataMap.get("dc:description", String.class);
-        } else if (metaDataMap != null && metaDataMap.containsKey("description")) {
-            description = metaDataMap.get("description", String.class);
-        } else if (valueMap.containsKey(JcrConstants.JCR_DESCRIPTION)) {
-            description = valueMap.get(JcrConstants.JCR_DESCRIPTION, String.class);
-        }
-
-        if (metaDataMap != null && metaDataMap.containsKey("allowedRenderFormat")) {
-            if (metaDataMap.get("allowedRenderFormat").equals("HTML")) {
-                tooltip = htmlTooltip;
-            } else {
-                tooltip = pdfTooltip;
-            }
-        }
-
-        if (resolver.getResource(thubmnail) == null) {
-            // thumbnail doesn't exist, set it to null
-            thubmnail = null;
-        }
-
-        return new Item(title, description, tooltip, path, thubmnail);
-    }
-
-    private class Item implements PortalLister.Item {
+    public static class Item implements PortalLister.Item {
         private String title;
         private String description;
         private String tooltip;
@@ -433,9 +165,5 @@ public class PortalListerImpl extends AbstractComponentImpl implements PortalLis
         public String getThumbnailLink() {
             return formThumbnail;
         }
-    }
-
-    private interface QueryStrategy {
-        void buildQuery(RequestParameter[] params, int predicateID, Map<String, String> queryMap);
     }
 }
