+++ bundles/core/src/test/resources/searchlister/searchlister-v1-withResults.json
@@ -1,0 +1,20 @@
+{
+    "id"                    : "searchlister-fb0b07e803",
+    "title"                 : "Sample Title",
+    "layout"                : "Custom",
+    "searchDisabled"        : true,
+    "sortDisabled"          : true,
+    "limit"                 : 4,
+    "searchResults"         : {
+        "data":[
+            {
+                "title":"Sample Form",
+                "description":"Sample description for Sample Form",
+                "tooltip":"Click here to view as HTML",
+                "formLink":"/content/dam/formsanddocuments/sample-form/jcr:content?wcmmode=disabled",
+                "thumbnailLink": null
+            }
+        ], "offset":0, "nextOffset":-1
+    },
+    ":type"                 : "core/fd/components/formsportal/searchlister/v1/searchlister"
+}
